Peek-A-Boo v1
This is a Jump List tester.  the jump list a system in Windows that tracks files access by a user.  this can be used for indirect evidence when performing forensics.  For example, a suspect says they never opened X file, the investigator can check the jump list and see that suspect did in fact open  said file. 
This PoSH GUI is used to check each Window OS when a new version becomes available to see if the jump list max value has changed.